@@ -15,14 +15,14 @@
 		"@skeletonlabs/tw-plugin": "^0.4.0",
 		"@sveltejs/adapter-auto": "^3.3.0",
 		"@sveltejs/kit": "^2.7.2",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@tailwindcss/forms": "^0.5.9",
 		"@types/node": "^22.7.9",
 		"@types/ol-ext": "npm:@siedlerchr/types-ol-ext",
 		"@vincjo/datatables": "^1.14.10",
 		"autoprefixer": "^10.4.20",
 		"postcss": "^8.4.47",
-		"svelte": "^4.2.19",
+		"svelte": "^5.0.0",
 		"svelte-check": "^4.0.5",
 		"typescript": "^5.6.3",
 		"vite": "^5.4.10"
