@@ -1,17 +1,27 @@
+<script lang="ts">
+    interface Props {
+        archiveSidebarLeft?: import('svelte').Snippet;
+        archivePageHeader?: import('svelte').Snippet;
+        children?: import('svelte').Snippet;
+    }
+
+    let { archiveSidebarLeft, archivePageHeader, children }: Props = $props();
+</script>
+
 <div id="archive-shell" data-testid="archive-shell" class="grid grid-cols-10 h-full">
     <div class="col-span-3">
-        <slot name="archiveSidebarLeft" />
+        {@render archiveSidebarLeft?.()}
     </div>
 
     <!-- Page -->
     <div id="archive-page" class="col-span-7">
         <!-- Slot: Page Header -->
-        {#if $$slots.archivePageHeader}
-            <header id="archive-page-header" class="flex-none"><slot name="archivePageHeader">(slot:archiveHeader)</slot></header>
+        {#if archivePageHeader}
+            <header id="archive-page-header" class="flex-none">{#if archivePageHeader}{@render archivePageHeader()}{:else}(slot:archiveHeader){/if}</header>
         {/if}
 
         <!-- Slot: Page Content (default) -->
-        <div id="archive-page-content" class="flex-auto w-full h-full flex overflow-hidden"><slot /></div>
+        <div id="archive-page-content" class="flex-auto w-full h-full flex overflow-hidden">{@render children?.()}</div>
     </div>
 </div>
 <style>
