@@ -1,7 +1,7 @@
 <script lang="ts">
     import Datatable from './Datatable.svelte';
     import ExecuteFinder from "./ExecuteFinder.svelte";
-    export let finderData;
+    let { finderData } = $props();
 </script>
 
 <ExecuteFinder finderData={finderData}></ExecuteFinder>
