@@ -1,15 +1,22 @@
 <script lang="ts">
     import ArchiveShell from '$lib/components/Archive/ArchiveShell.svelte';
 	import type { LayoutData } from './$types';
-	export let data: LayoutData;
     import { page } from '$app/stores';  
     import Sidebar from '$lib/components/Archive/Detail/Sidebar/Sidebar.svelte';
+    interface Props {
+        data: LayoutData;
+        children?: import('svelte').Snippet;
+    }
+
+    let { data, children }: Props = $props();
 
 </script>
 
 <ArchiveShell>
-    <svelte:fragment slot="archiveSidebarLeft">
-        <Sidebar finderData={data.finderData}/>
-    </svelte:fragment>
-    <slot /> 
+    {#snippet archiveSidebarLeft()}
+    
+            <Sidebar finderData={data.finderData}/>
+        
+    {/snippet}
+    {@render children?.()} 
 </ArchiveShell>
