@@ -1,1 +1,9 @@
-<slot />
+<script lang="ts">
+	interface Props {
+		children?: import('svelte').Snippet;
+	}
+
+	let { children }: Props = $props();
+</script>
+
+{@render children?.()}
