@@ -2,13 +2,20 @@
     import ImagerySideBar from '$lib/components/Imagery/ImagerySideBar.svelte';
     import ImageryShell from '$lib/components/Imagery/ImageryShell.svelte';
 	import type { LayoutData } from './$types';
-	export let data: LayoutData;
+    interface Props {
+        data: LayoutData;
+        children?: import('svelte').Snippet;
+    }
+
+    let { data, children }: Props = $props();
 </script>
 
 
 <ImageryShell>
-    <svelte:fragment slot="imagerySidebarLeft">
-        <ImagerySideBar images={data.images}  />
-    </svelte:fragment>
-    <slot /> 
+    {#snippet imagerySidebarLeft()}
+    
+            <ImagerySideBar images={data.images}  />
+        
+    {/snippet}
+    {@render children?.()} 
 </ImageryShell>
